Updated the description of the project
Made an update on the README.md file for the master branch. The
description of the project gives a step wise process
of what version control is and how the project was carried out.
This description primarily helps to gives an insight or a better
understanding of what the project is about.